Re-orginaized the functions associted with "startup" into a new vcastInstallation file, and rearagned a few other functions to clean up the importing treee.
