+++ src/client.ts
@@ -1,7 +1,7 @@
 // this file contains the language server client logic for test script editing
 
 import * as path from "path";
-import { workspace, ExtensionContext } from "vscode";
+import { ExtensionContext } from "vscode";
 
 import {
   LanguageClient,
@@ -9,8 +9,10 @@ import {
   ServerOptions,
   TransportKind,
 } from "vscode-languageclient";
+
 import { vectorMessage } from "./messagePane";
-import { vPythonCommandToUse } from "./utilities";
+
+import { vPythonCommandToUse } from "./vcastInstallation";
 
 let client: LanguageClient;
