+++ src/vcastInstallation.ts
@@ -1,0 +1,444 @@
+import * as vscode from "vscode";
+
+import { showSettings } from "./helper";
+
+import { openMessagePane, vectorMessage } from "./messagePane";
+
+import {
+  executeCommandSync,
+  exeFilename,
+  executeVPythonScript,
+} from "./utilities";
+
+import { vcastLicenseOK } from "./vcastAdapter";
+
+const fs = require("fs");
+const path = require("path");
+const which = require("which");
+
+export const clicastName = "clicast";
+export let clicastCommandToUse: string;
+
+const vPythonName = "vpython";
+const vpythonFromPath = which.sync(vPythonName, { nothrow: true });
+export let vPythonCommandToUse: string;
+
+const vcastqtName = "vcastqt";
+export let vcastCommandToUse: string | undefined = undefined;
+
+let globalCheckSumCommand: string | undefined = undefined;
+let crc32Name = "crc32-win32.exe";
+if (process.platform == "linux") {
+  crc32Name = "crc32-linux";
+}
+
+export let globalTestInterfacePath: string | undefined = undefined;
+let globalCrc32Path: string | undefined = undefined;
+
+export let globalPathToSupportFiles: string;
+
+let globalCodedTestingAvailable: boolean = false;
+
+export let globalIncludePath: string | undefined = undefined;
+
+const atgName = "atg";
+export let atgCommandToUse: string | undefined = undefined;
+export let atgAvailable: boolean = false;
+
+export const configurationFile = "c_cpp_properties.json";
+export const launchFile = "launch.json";
+
+// ---------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------
+export function initializeInstallerFiles(context: vscode.ExtensionContext) {
+  const pathToTestInterface = context.asAbsolutePath(
+    "./python/vTestInterface.py"
+  );
+  if (fs.existsSync(pathToTestInterface)) {
+    vectorMessage("Found vTestInterface here: " + pathToTestInterface);
+    globalTestInterfacePath = `${pathToTestInterface}`;
+  }
+
+  const crc32Path = context.asAbsolutePath("./python/crc32.py");
+  if (fs.existsSync(crc32Path)) {
+    vectorMessage("Found the crc32 python wrapper here: " + crc32Path);
+    globalCrc32Path = `${crc32Path}`;
+  }
+
+  const pathToSupportFiles = context.asAbsolutePath("./supportFiles");
+  if (fs.existsSync(pathToSupportFiles)) {
+    vectorMessage("Found extension support files here: " + pathToSupportFiles);
+    globalPathToSupportFiles = `${pathToSupportFiles}`;
+  }
+}
+
+function vcastVersionGreaterThan(
+  vcastInstallationPath: string,
+  version: number,
+  servicePack: number
+): boolean {
+  // A general purpose version checker, will be needed for Coded Tests, etc.
+
+  let returnValue = false;
+  const toolPath = path.join(vcastInstallationPath, "DATA", "tool_version.txt");
+
+  const toolVersion = fs.readFileSync(toolPath).toString().trim();
+  // extract version and service pack from toolVersion (23.sp2 date)
+  const matched = toolVersion.match(/(\d+)\.sp(\d+).*/);
+  if (matched) {
+    const tooVersion = parseInt(matched[1]);
+    const toolServicePack = parseInt(matched[2]);
+    if (
+      tooVersion > version ||
+      (tooVersion == version && toolServicePack >= servicePack)
+    )
+      returnValue = true;
+  }
+  // this allows us to work with development builds for internal testing
+  else if (toolVersion.includes(" revision ")) {
+    returnValue = true;
+  }
+  return returnValue;
+}
+
+function vectorCASTSupportsATG(vcastInstallationPath: string): boolean {
+  // Versions of VectorCAST between 23sp0 and 23sp4 had ATG but since
+  // we changed the ATG command line interface with 23sp5, we have decided
+  // to only support versions greater than that.
+
+  return vcastVersionGreaterThan(vcastInstallationPath, 23, 5);
+}
+
+function checkForATG(vcastInstallationPath: string) {
+  // we only set atgCommandToUse if we find atg and it's licensed
+  const atgCommand = path.join(vcastInstallationPath, exeFilename(atgName));
+  let statusMessageText = "";
+  if (fs.existsSync(atgCommand)) {
+    statusMessageText = `   found '${atgName}' here: ${vcastInstallationPath}`;
+    const candidateCommand = atgCommand;
+
+    // now check if its licensed ... just atg --help and check the exit code
+    const commandToRun: string = `${candidateCommand} --help`;
+
+    // cwd=working dir for this process /  printErrorDetails=false
+    const commandStatus = executeCommandSync(
+      commandToRun,
+      process.cwd(),
+      false
+    );
+    if (commandStatus.errorCode == 0) {
+      statusMessageText += ", license is available";
+      atgCommandToUse = candidateCommand;
+    } else {
+      statusMessageText += ", license is NOT available";
+    }
+    vectorMessage(statusMessageText);
+    atgAvailable =
+      atgCommandToUse != undefined &&
+      vectorCASTSupportsATG(vcastInstallationPath);
+
+    // atgAvailabel is used by package.json to control the existance of the atg command in the context menus
+    vscode.commands.executeCommand(
+      "setContext",
+      "vectorcastTestExplorer.atgAvailable",
+      atgAvailable
+    );
+  } else {
+    vectorMessage(
+      `   could NOT find '${atgName}' here: ${vcastInstallationPath}`
+    );
+  }
+}
+
+function findVcastTools(): boolean {
+  // This function will set global paths to vpython, clicast and vcastqt
+  // by sequentially looking for vpython in the directory set via the
+  // 1. extension option: "vectorcastInstallationLocation"
+  // 2. VECTORCAST_DIR
+  // 3. system PATH variable
+
+  // return value
+  let foundAllvcastTools = false;
+
+  // value of the extension option
+  const settings = vscode.workspace.getConfiguration("vectorcastTestExplorer");
+  const installationOptionString = settings.get(
+    "vectorcastInstallationLocation",
+    ""
+  );
+
+  // value of VECTORCAST_DIR
+  const VECTORCAST_DIR = process.env["VECTORCAST_DIR"];
+
+  // VectorCAST installation location
+  let vcastInstallationPath: string | undefined = undefined;
+
+  // priority 1 is the option value, since this lets the user over-ride PATH or VECTORCAST_DIR
+  if (installationOptionString.length > 0) {
+    const candidatePath = path.join(
+      installationOptionString,
+      exeFilename(vPythonName)
+    );
+    if (fs.existsSync(candidatePath)) {
+      vcastInstallationPath = installationOptionString;
+      vPythonCommandToUse = candidatePath;
+      vectorMessage(
+        `   found '${vPythonName}' using the 'Vectorcast Installation Location' option [${installationOptionString}].`
+      );
+    } else {
+      vectorMessage(
+        `   the installation path provided: '${installationOptionString}' does not contain ${vPythonName}, ` +
+          "use the extension options to provide a valid VectorCAST installation directory."
+      );
+      showSettings();
+    }
+  }
+
+  // priority 2 is VECTORCAST_DIR, while this is no longer required, it is still widely used
+  else if (VECTORCAST_DIR) {
+    const candidatePath = path.join(VECTORCAST_DIR, exeFilename(vPythonName));
+    if (fs.existsSync(candidatePath)) {
+      vcastInstallationPath = VECTORCAST_DIR;
+      vPythonCommandToUse = candidatePath;
+      vectorMessage(
+        `   found '${vPythonName}' using VECTORCAST_DIR [${VECTORCAST_DIR}]`
+      );
+    } else {
+      vectorMessage(
+        `   the installation path provided via VECTORCAST_DIR does not contain ${vPythonName}, ` +
+          "use the extension options to provide a valid VectorCAST installation directory."
+      );
+      showSettings();
+    }
+  }
+
+  // priority 3 is the system path
+  else if (vpythonFromPath) {
+    vcastInstallationPath = path.dirname(vpythonFromPath);
+    vPythonCommandToUse = vpythonFromPath;
+    vectorMessage(
+      `   found '${vPythonName}' on the system path [${vcastInstallationPath}]`
+    );
+  } else {
+    vectorMessage(
+      `   command: '${vPythonName}' is not on the system PATH, and VECTORCAST_DIR is not set, ` +
+        "use the extension options to provide a valid VectorCAST installation directory."
+    );
+    showSettings();
+  }
+
+  // if we found a vpython somewhere ...
+  // we assume the other executables are there too,  but we check anyway :)
+  if (vcastInstallationPath) {
+    // do all of the setup required to use clicast
+    foundAllvcastTools = initializeVcastUtilities(vcastInstallationPath);
+
+    // setup coded-test stuff (new for vc24)
+    initializeCodedTestSupport(vcastInstallationPath);
+
+    // check if we have access to a valid crc32 command - this is not fatal
+    // must be called after initializeInstallerFiles()
+
+    if (!initializeChecksumCommand(vcastInstallationPath)) {
+      vscode.window.showWarningMessage(
+        "The VectorCAST Test Explorer could not find the required VectorCAST CRC-32 module, " +
+          "so the code coverage feature will not be available.  For details on how to resolve " +
+          "this issue, please refer to the 'Prerequisites' section of the README.md file."
+      );
+    }
+  }
+
+  return foundAllvcastTools;
+}
+
+export function checkIfInstallationIsOK() {
+  // Check if the installation is ok by verifying that:
+  //   - we can find vpython and clicast
+  //   - we have a valid license
+
+  // default this to false, it only gets to true if we find
+  // vpython and clicast and we have a license
+  let returnValue = false;
+
+  vectorMessage("Checking that a VectorCAST installation is available ... ");
+
+  if (findVcastTools()) {
+    // check if we have a valid license
+
+    if (vcastLicenseOK()) {
+      vectorMessage("   VectorCAST license is available ...");
+      returnValue = true;
+    } else {
+      vectorMessage("   no VectorCAST license is available");
+      returnValue = false;
+    }
+  }
+
+  vectorMessage("-".repeat(100) + "\n");
+
+  if (!returnValue) {
+    vectorMessage(
+      "Please refer to the installation and configuration instructions for details on resolving these issues"
+    );
+    openMessagePane();
+  }
+  return returnValue;
+}
+
+function initializeVcastUtilities(vcastInstallationPath: string) {
+  let toolsFound = false;
+  clicastCommandToUse = path.join(
+    vcastInstallationPath,
+    exeFilename(clicastName)
+  );
+
+  if (fs.existsSync(clicastCommandToUse)) {
+    vectorMessage(`   found '${clicastName}' here: ${vcastInstallationPath}`);
+    vcastCommandToUse = path.join(
+      vcastInstallationPath,
+      exeFilename(vcastqtName)
+    );
+    if (fs.existsSync(vcastCommandToUse)) {
+      vectorMessage(`   found '${vcastqtName}' here: ${vcastInstallationPath}`);
+
+      // we only set toolsFound if we find clicast AND vcastqt
+      toolsFound = true;
+
+      // atg existing or being licensed does NOT affect toolsFound
+      checkForATG(vcastInstallationPath);
+    } else {
+      vectorMessage(
+        `   could NOT find '${vcastqtName}' here: ${vcastInstallationPath}`
+      );
+    }
+  } else {
+    vectorMessage(
+      `   could NOT find '${clicastName}' here: ${vcastInstallationPath}`
+    );
+  }
+  return toolsFound;
+}
+
+function pyCrc32IsAvailable(): boolean {
+  // Although crc32.py simply puts out AVAILABLE or NOT-AVAILABLE,
+  // vpython prints this annoying message if VECTORCAST_DIR does not match the executable
+  // so we need this logic to handle that case.
+
+  // I'm doing this in multiple steps for clarity
+  const commandOutputText = executeVPythonScript(
+    `${vPythonCommandToUse} ${globalCrc32Path}`,
+    process.cwd()
+  ).stdout;
+  const outputLinesAsArray = commandOutputText.split("\n");
+  const lastOutputLine = outputLinesAsArray[outputLinesAsArray.length - 1];
+  return lastOutputLine == "AVAILABLE";
+}
+
+function getCRCutilityPath(vcastInstallationPath: string) {
+  // check if the crc32 utility has been added to the the VectorCAST installation
+
+  let returnValue: string | undefined = undefined;
+  if (vcastInstallationPath) {
+    let candidatePath = path.join(vcastInstallationPath, crc32Name);
+    if (fs.existsSync(candidatePath)) {
+      vectorMessage(`   found '${crc32Name}' here: ${vcastInstallationPath}`);
+      returnValue = candidatePath;
+    } else {
+      vectorMessage(
+        `   could NOT find '${crc32Name}' here: ${vcastInstallationPath}, coverage annotations will not be available`
+      );
+    }
+  }
+  return returnValue;
+}
+
+function initializeChecksumCommand(
+  vcastInstallationPath: string
+): string | undefined {
+  // checks if this vcast distro has python checksum support built-in
+  if (globalCrc32Path && pyCrc32IsAvailable()) {
+    globalCheckSumCommand = `${vPythonCommandToUse} ${globalCrc32Path}`;
+  } else {
+    // check if the user has patched the distro with the crc32 utility
+    globalCheckSumCommand = getCRCutilityPath(vcastInstallationPath);
+  }
+
+  return globalCheckSumCommand;
+}
+
+export function getChecksumCommand() {
+  return globalCheckSumCommand;
+}
+
+function shouldPromptForIncludePath(includePath: string): boolean {
+  // So that we don't annoy users with the coded-test popup every time,
+  // we do our best to check if the include path is already in the workspace
+
+  let returnValue: boolean = true;
+  if (fs.existsSync(includePath)) {
+    for (const workspace of vscode.workspace.workspaceFolders || []) {
+      const workspaceRoot = workspace.uri.fsPath;
+      const c_cpp_properties = path.join(
+        workspaceRoot,
+        ".vscode",
+        configurationFile
+      );
+      if (fs.existsSync(c_cpp_properties)) {
+        const c_cpp_properties_contents = fs
+          .readFileSync(c_cpp_properties)
+          .toString();
+        if (c_cpp_properties_contents.includes(includePath)) {
+          returnValue = false;
+          break;
+        }
+      }
+    }
+  } else {
+    // don't prompt if the include path is missing for some reason
+    returnValue = false;
+  }
+  return returnValue;
+}
+
+function checkWorkspaceForIncludePath(includePath: string) {
+  // We'd like to make it easy for the user to add the include path
+  // for the VectorCAST vUnit/Include directory.  I looked at automating
+  // this via the vscode.workspace.getConfiguration() API,
+  // but there are too many edge cases, I decided to do check
+  // if the path exists, in any of the c_cpp_properties.json files
+  // and prompt the user to add if it doesn't
+
+  // swap backslashes to make paths consistent for windows users and
+  // so that they can copy paste from the pop-up to the .json
+  globalIncludePath = includePath.replace(/\\/g, "/");
+
+  if (shouldPromptForIncludePath(globalIncludePath)) {
+    vscode.window.showInformationMessage(
+      "The include path for VectorCAST Coded Testing was not found in your workspace, you should add the " +
+        `include path by right clicking on the appropriate ${configurationFile} file, ` +
+        "and choosing 'VectorCAST: Add Coded Test Include Path`  "
+    );
+  }
+}
+
+function initializeCodedTestSupport(vcastInstallationPath: string) {
+  // When we get here vcastInstallationPath will point to a
+  // valid VectorCAST installation but we don't know if
+  // this version has coded test support, so check for that
+  // and initialize global variables to support coded testing
+
+  const candidatePath = path.join(vcastInstallationPath, "vunit", "include");
+  if (fs.existsSync(candidatePath)) {
+    vectorMessage(`   found coded-test support, initializing ...`);
+    globalCodedTestingAvailable = true;
+    checkWorkspaceForIncludePath(candidatePath);
+  } else {
+    globalCodedTestingAvailable = false;
+  }
+  // this controls the availability of the Add Coded Test Include Path context menu item
+  vscode.commands.executeCommand(
+    "setContext",
+    "vectorcastTestExplorer.codedTestingAvailable",
+    globalCodedTestingAvailable
+  );
+}
