+++ src/vcastUtilities.ts
@@ -5,19 +5,16 @@ import * as jsonc from "jsonc-parser";
 
 import { openMessagePane, vectorMessage } from "./messagePane";
 
-import { compoundOnlyString, getTestNode, testNodeType } from "./testData";
+import { getTestNode, testNodeType } from "./testData";
 
 import { updateTestPane } from "./testPane";
 
 import {
   commandStatusType,
-  executeCommandSync,
-  exeFilename,
   jsoncModificationOptions,
   jsoncParseErrors,
   jsoncParseOptions,
   openFileWithLineSelected,
-  processExceptionFromExecuteCommand,
 } from "./utilities";
 
 import {
@@ -27,136 +24,18 @@ import {
   runBasisPathCommands,
 } from "./vcastAdapter";
 
+import {
+  clicastCommandToUse,
+  globalIncludePath,
+  globalTestInterfacePath,
+  vPythonCommandToUse,
+} from "./vcastInstallation";
+
 const fs = require("fs");
 const os = require("os");
 const path = require("path");
-const spawn = require("child_process").spawn;
 
-export const configurationFile = "c_cpp_properties.json";
-export const launchFile = "launch.json";
 
-const clicastName = "clicast";
-export let clicastCommandToUse: string;
-
-const vcastqtName = "vcastqt";
-export let vcastCommandtoUse: string | undefined = undefined;
-
-const atgName = "atg";
-export let atgCommandToUse: string | undefined = undefined;
-export let atgAvailable: boolean = false;
-
-export let codedTestAvailable: boolean = false;
-
-function vcastVersionGreaterThan(
-  vcastInstallationPath: string,
-  version: number,
-  servicePack: number
-): boolean {
-  // A general purpose version checker, will be needed for Coded Tests, etc.
-
-  let returnValue = false;
-  const toolPath = path.join(vcastInstallationPath, "DATA", "tool_version.txt");
-
-  const toolVersion = fs.readFileSync(toolPath).toString().trim();
-  // extract version and service pack from toolVersion (23.sp2 date)
-  const matched = toolVersion.match(/(\d+)\.sp(\d+).*/);
-  if (matched) {
-    const tooVersion = parseInt(matched[1]);
-    const toolServicePack = parseInt(matched[2]);
-    if (
-      tooVersion > version ||
-      (tooVersion == version && toolServicePack >= servicePack)
-    )
-      returnValue = true;
-  }
-  // this allows us to work with development builds for internal testing
-  else if (toolVersion.includes(" revision ")) {
-    returnValue = true;
-  }
-  return returnValue;
-}
-
-function vectorCASTSupportsATG(vcastInstallationPath: string): boolean {
-  // Versions of VectorCAST between 23sp0 and 23sp4 had ATG but since
-  // we changed the ATG command line interface with 23sp5, we have decided
-  // to only support versions greater than that.
-
-  return vcastVersionGreaterThan(vcastInstallationPath, 23, 5);
-}
-
-function checkForATG(vcastInstallationPath: string) {
-  // we only set atgCommandToUse if we find atg and it's licensed
-  const atgCommand = path.join(vcastInstallationPath, exeFilename(atgName));
-  let statusMessageText = "";
-  if (fs.existsSync(atgCommand)) {
-    statusMessageText = `   found '${atgName}' here: ${vcastInstallationPath}`;
-    const candidateCommand = atgCommand;
-
-    // now check if its licensed ... just atg --help and check the exit code
-    const commandToRun: string = `${candidateCommand} --help`;
-
-    // cwd=working dir for this process /  printErrorDetails=false
-    const commandStatus = executeCommandSync(
-      commandToRun,
-      process.cwd(),
-      false
-    );
-    if (commandStatus.errorCode == 0) {
-      statusMessageText += ", license is available";
-      atgCommandToUse = candidateCommand;
-    } else {
-      statusMessageText += ", license is NOT available";
-    }
-    vectorMessage(statusMessageText);
-    atgAvailable =
-      atgCommandToUse != undefined &&
-      vectorCASTSupportsATG(vcastInstallationPath);
-
-    // atgAvailabel is used by package.json to control the existance of the atg command in the context menus
-    vscode.commands.executeCommand(
-      "setContext",
-      "vectorcastTestExplorer.atgAvailable",
-      atgAvailable
-    );
-  } else {
-    vectorMessage(
-      `   could NOT find '${atgName}' here: ${vcastInstallationPath}`
-    );
-  }
-}
-
-function shouldPromptForIncludePath(includePath: string): boolean {
-  // So that we don't annoy users with the coded-test popup every time,
-  // we do our best to check if the include path is already in the workspace
-
-  let returnValue: boolean = true;
-  if (fs.existsSync(includePath)) {
-    for (const workspace of vscode.workspace.workspaceFolders || []) {
-      const workspaceRoot = workspace.uri.fsPath;
-      const c_cpp_properties = path.join(
-        workspaceRoot,
-        ".vscode",
-        configurationFile
-      );
-      if (fs.existsSync(c_cpp_properties)) {
-        const c_cpp_properties_contents = fs
-          .readFileSync(c_cpp_properties)
-          .toString();
-        if (c_cpp_properties_contents.includes(includePath)) {
-          returnValue = false;
-          break;
-        }
-      }
-    }
-  } else {
-    // don't prompt if the include path is missing for some reason
-    returnValue = false;
-  }
-  return returnValue;
-}
-
-let globalIncludePath: string | undefined = undefined;
-let globalCodedTestingAvailable: boolean = false;
 export function addIncludePath(fileUri: vscode.Uri) {
   // I'm handling a few error cases here without going crazy
   let statusMessages: string[] = [];
@@ -243,83 +122,6 @@ export function addIncludePath(fileUri: vscode.Uri) {
 
   // we unconditionally write rather than tracking if we changed anything
   fs.writeFileSync(fileUri.fsPath, existingJSONasString);
-}
-
-export function initializeCodedTestSupport(vcastInstallationPath: string) {
-  // When we get here vcastInstallationPath will point to a
-  // valid VectorCAST installation but we don't know if
-  // this version has coded test support, so check for that
-  // and initialize global variables to support coded testing
-
-  const candidatePath = path.join(vcastInstallationPath, "vunit", "include");
-  if (fs.existsSync(candidatePath)) {
-    vectorMessage(`   found coded-test support, initializing ...`);
-    globalCodedTestingAvailable = true;
-    checkWorkspaceForIncludePath(candidatePath);
-  } else {
-    globalCodedTestingAvailable = false;
-  }
-  // this controls the availability of the Add Coded Test Include Path context menu item
-  vscode.commands.executeCommand(
-    "setContext",
-    "vectorcastTestExplorer.codedTestingAvailable",
-    globalCodedTestingAvailable
-  );
-}
-
-function checkWorkspaceForIncludePath(includePath: string) {
-  // We'd like to make it easy for the user to add the include path
-  // for the VectorCAST vUnit/Include directory.  I looked at automating
-  // this via the vscode.workspace.getConfiguration() API,
-  // but there are too many edge cases, I decided to do check
-  // if the path exists, in any of the c_cpp_properties.json files
-  // and prompt the user to add if it doesn't
-
-  // swap backslashes to make paths consistent for windows users and
-  // so that they can copy paste from the pop-up to the .json
-  globalIncludePath = includePath.replace(/\\/g, "/");
-
-  if (shouldPromptForIncludePath(globalIncludePath)) {
-    vscode.window.showInformationMessage(
-      "The include path for VectorCAST Coded Testing was not found in your workspace, you should add the " +
-        `include path by right clicking on the appropriate ${configurationFile} file, ` +
-        "and choosing 'VectorCAST: Add Coded Test Include Path`  "
-    );
-  }
-}
-
-export function initializeVcastUtilities(vcastInstallationPath: string) {
-  let toolsFound = false;
-  clicastCommandToUse = path.join(
-    vcastInstallationPath,
-    exeFilename(clicastName)
-  );
-
-  if (fs.existsSync(clicastCommandToUse)) {
-    vectorMessage(`   found '${clicastName}' here: ${vcastInstallationPath}`);
-    vcastCommandtoUse = path.join(
-      vcastInstallationPath,
-      exeFilename(vcastqtName)
-    );
-    if (fs.existsSync(vcastCommandtoUse)) {
-      vectorMessage(`   found '${vcastqtName}' here: ${vcastInstallationPath}`);
-
-      // we only set toolsFound if we find clicast AND vcastqt
-      toolsFound = true;
-
-      // atg existing or being licensed does NOT affect toolsFound
-      checkForATG(vcastInstallationPath);
-    } else {
-      vectorMessage(
-        `   could NOT find '${vcastqtName}' here: ${vcastInstallationPath}`
-      );
-    }
-  } else {
-    vectorMessage(
-      `   could NOT find '${clicastName}' here: ${vcastInstallationPath}`
-    );
-  }
-  return toolsFound;
 }
 
 function convertTestScriptContents(scriptPath: string) {
@@ -414,8 +216,7 @@ export function generateAndLoadBasisPathTests(testNode: testNodeType) {
   // ignore the testName (if any)
   testNode.testName = "";
 
-  runBasisPathCommands(testNode, tempScriptPath);
-  
+  runBasisPathCommands(testNode, tempScriptPath, loadScriptCallBack);
 }
 
 export function generateAndLoadATGTests(testNode: testNodeType) {
@@ -439,10 +240,10 @@ export function generateAndLoadATGTests(testNode: testNodeType) {
   // ignore the testName (if any)
   testNode.testName = "";
 
-  runATGCommands(testNode, tempScriptPath);
+  runATGCommands(testNode, tempScriptPath, loadScriptCallBack);
 }
 
-export function loadScriptCallBack(
+function loadScriptCallBack(
   commandStatus: commandStatusType,
   enviroName: string,
   scriptPath: string
@@ -467,193 +268,6 @@ export function loadScriptCallBack(
     vectorMessage(commandStatus.stdout);
     openMessagePane();
   }
-}
-
-export function executeWithRealTimeEcho(
-  command: string,
-  argList: string[],
-  CWD: string,
-  callback?: any,
-  enviroPath?: string
-) {
-  // this function is used to build and rebuild environments
-  // long running commands where we want to show real-time output
-
-  // it uses spawn to execute a clicast command, log the output to the
-  // message pane, and update the test explorer when the command completes
-
-  // To debug what's going on with vcast, you can add -dall to
-  // argList, which will dump debug info for the clicast invocation
-  let clicast = spawn(command, argList, { cwd: CWD });
-  vectorMessage("-".repeat(100));
-
-  // maybe this is a hack, but after reading stackoverflow for a while I could
-  // not come up with anything better.  The issue is that the on ("exit") gets called
-  // before the stdout stream is closed so stdoutBuffer is incomplete at that point
-  // so we use on ("exit") to invoke the callback and on ("close") to dump the clicast stdout.
-
-  // I tried to only dump the output when the exit code was non 0 but we get a race
-  // condition because the exit might not have saved it when the close is seen.
-
-  vectorMessage("-".repeat(100));
-  clicast.stdout.on("data", function (data: any) {
-    // split raw message based on \n or \r because messages
-    // that come directly from the compiler are LF terminated
-    const rawString = data.toString();
-    const lineArray = rawString.split(/[\n\r?]/);
-    for (const line of lineArray) {
-      if (line.length > 0) {
-        vectorMessage(line.replace(/\n/g, ""));
-      }
-    }
-  });
-
-  clicast.stdout.on("close", function (code: any) {
-    vectorMessage("-".repeat(100));
-  });
-
-  clicast.on("exit", function (code: any) {
-    vectorMessage("-".repeat(100));
-    vectorMessage(
-      `${clicastName}: '${argList.join(
-        " "
-      )}' returned exit code: ${code.toString()}`
-    );
-    vectorMessage("-".repeat(100));
-    if (callback) {
-      callback(enviroPath, code);
-    }
-  });
-}
-
-interface statusMessageType {
-  fullLines: string;
-  remainderText: string;
-}
-function processCommandOutput(
-  remainderTextFromLastCall: string,
-  newTextFromThisCall: string
-): statusMessageType {
-  // The purpose of this function is to process the raw text that comes
-  // from the spawned process and to split it into full lines and a "remainder"
-  // The caller will keep the remainder around until the next data comes in
-  // and then pass that in with the new text.
-
-  let returnObject: statusMessageType = { fullLines: "", remainderText: "" };
-  const candidateString = remainderTextFromLastCall + newTextFromThisCall;
-
-  if (candidateString.endsWith("\n"))
-    // if we got all full lines, there is no remainder
-    returnObject.fullLines = candidateString.slice(
-      0,
-      candidateString.length - 1
-    );
-  else if (candidateString.includes("\n")) {
-    // if there is at least one \n then we have full lines and a remainder
-    const whereToSplit = candidateString.lastIndexOf("\n");
-    returnObject.fullLines = candidateString.substring(0, whereToSplit);
-    returnObject.remainderText = candidateString.substring(
-      whereToSplit + 1,
-      candidateString.length
-    );
-  } else {
-    // otherwise we have only a remainder
-    returnObject.remainderText = candidateString;
-  }
-
-  return returnObject;
-}
-
-export function executeClicastWithProgress(
-  title: string,
-  commandAndArgs: string[],
-  enviroName: string,
-  testScriptPath: string,
-  filter: RegExp,
-  callback: any
-) {
-  // Very similar to the executeWithRealTimeEcho(), but adds a progress dialog,
-  // and a different callback structure.
-  // We use this for generating the basis path and ATG tests (for now)
-
-  vectorMessage(`Executing command: ${commandAndArgs.join(" ")}`);
-  let commandStatus: commandStatusType = { errorCode: 0, stdout: "" };
-
-  const cwd = path.dirname(testScriptPath);
-  const command = commandAndArgs[0];
-  const args = commandAndArgs.slice(1, commandAndArgs.length);
-
-  vscode.window.withProgress(
-    {
-      location: vscode.ProgressLocation.Notification,
-      title: title,
-      cancellable: false,
-    },
-    async (progress) => {
-      return new Promise(async (resolve, reject) => {
-        // shell is needed so that stdout is NOT buffered
-        const commandHandle = spawn(command, args, { cwd: cwd, shell: true });
-
-        // each time we get an entry here, we need to check if we have a
-        // partial message if so we print the part up the the
-        // final \n and buffer the rest, see comment above
-        let remainderTextFromLastCall = "";
-
-        commandHandle.stdout.on("data", async (data: any) => {
-          const message: statusMessageType = processCommandOutput(
-            remainderTextFromLastCall,
-            data.toString()
-          );
-          remainderTextFromLastCall = message.remainderText;
-
-          if (message.fullLines.length > 0) {
-            vectorMessage(message.fullLines);
-
-            // for the dialog, we want use the filter to decide what to show
-            // and this requires the message data to be split into single lines
-            const lineArray = message.fullLines.split("\n");
-            for (const line of lineArray) {
-              const matched = line.match(filter);
-              if (matched && matched.length > 0) {
-                progress.report({ message: matched[0], increment: 10 });
-                // This is needed to allow the message window to update ...
-                await new Promise<void>((r) => setTimeout(r, 0));
-              }
-            }
-          }
-        });
-
-        commandHandle.stderr.on("data", async (data: any) => {
-          const message: statusMessageType = processCommandOutput(
-            remainderTextFromLastCall,
-            data.toString(data)
-          );
-          remainderTextFromLastCall = message.remainderText;
-
-          if (message.fullLines.length > 0) {
-            vectorMessage(message.fullLines);
-          }
-        });
-
-        commandHandle.on("error", (error: any) => {
-          commandStatus = processExceptionFromExecuteCommand(
-            commandAndArgs.join(" "),
-            error,
-            true
-          );
-        });
-        commandHandle.on("close", (code: any) => {
-          // display any remaining text ...
-          if (remainderTextFromLastCall.length > 0) {
-            vectorMessage(remainderTextFromLastCall);
-          }
-          commandStatus.errorCode = code;
-          resolve(code);
-          callback(commandStatus, enviroName, testScriptPath);
-        });
-      });
-    }
-  );
 }
 
 export enum testStatus {
@@ -737,46 +351,61 @@ export function getEnviroNameFromFile(filePath: string): string | undefined {
   return enviroName;
 }
 
-function quote(name: string) {
-  // if name contains <<COMPOUND>>, <<INIT>> or parenthesis
-  // we need to quote the name so that the shell does not interpret it.
+export function testInterfaceCommand(
+  mode: string,
+  enviroPath: string,
+  testID: string = ""
+): any | undefined {
+  // enviroPath is the absolute path to the environnement directory
+  // testID is contains the string that uniquely identifies the node, something like:
+  //    vcast:TEST|manager.Manager::PlaceOrder.test-Manager::PlaceOrder
+  //    vcast:unitTests/MANAGER|manager.Manager::PlaceOrder.test-Manager::PlaceOrder
+  //
 
-  if (
-    name.includes("<") ||
-    name.includes(">") ||
-    name.includes("(") ||
-    name.includes(")")
-  ) {
-    return '"' + name + '"';
-  } else return name;
+  if (globalTestInterfacePath && vPythonCommandToUse) {
+    const command = `${vPythonCommandToUse} ${globalTestInterfacePath} --mode=${mode} --clicast=${clicastCommandToUse} --path=${enviroPath}`;
+    let testArgument = "";
+    if (testID.length > 0) {
+      // we need to strip the "path part" of the environment directory from the test ID
+      // which is the part before the '|' and after the ':'
+      const enviroPath = testID.split("|")[0].split(":")[1];
+
+      // now the path to the environment might have a slash if the environment is nested or not
+      // so we need to handle that case, since we only want the environment name
+      let enviroName = enviroPath;
+      if (enviroName.includes("/")) {
+        enviroName = enviroPath.substring(
+          enviroPath.lastIndexOf("/") + 1,
+          enviroPath.length
+        );
+      }
+      // The -test arguments should be the enviro name along with everything after the |
+      testArgument = ` --test="${enviroName}|${testID.split("|")[1]}"`;
+    }
+    return command + testArgument;
+  } else
+    vscode.window.showWarningMessage(
+      "The VectorCAST Test Explorer could not find the vpython utility."
+    );
+  return undefined;
 }
 
-export function getClicastArgsFromTestNodeAsList(
-  testNode: testNodeType
-): string[] {
-  // this function will create the enviro, unit, subprogram, and test
-  // arguments as a list, since spawn for example requires an arg list.
-
-  let returnList = [];
-  returnList.push(`-e${testNode.enviroName}`);
-  if (testNode.unitName.length > 0 && testNode.unitName != "not-used")
-    returnList.push(`-u${testNode.unitName}`);
-
-  // we need the quotes on the names to handle <<COMPOUND>>/<<INIT>>/parenthesis
-  if (testNode.functionName.length > 0)
-    returnList.push(`-s${quote(testNode.functionName)}`);
-  if (testNode.testName.length > 0) {
-    const nameToUse = testNode.testName.replace(compoundOnlyString, "");
-    returnList.push(`-t${quote(nameToUse)}`);
-  }
-
-  return returnList;
+export function parseCBTCommand(filePath: string): string {
+  // this command returns the list of tests that exist in a coded test source file
+  return `${vPythonCommandToUse} ${globalTestInterfacePath} --mode=parseCBT --path=${filePath}`;
 }
 
-export function getClicastArgsFromTestNode(testNode: testNodeType) {
-  // this function will create the enviro, unit, subprogram,
-  // and test arg string for clicast calls that need a arg string
+export function rebuildEnvironmentCommand(filePath: string): string {
+  // this command performs the environment rebuild, including the update of the .env file
 
-  const argList = getClicastArgsFromTestNodeAsList(testNode);
-  return argList.join(" ");
+  // read the settings that affect enviro build
+  const settings = vscode.workspace.getConfiguration("vectorcastTestExplorer");
+  var optionsDict: { [command: string]: string | boolean } = {};
+  optionsDict["ENVIRO.COVERAGE_TYPE"] = settings.get(
+    "build.coverageKind",
+    "None"
+  );
+
+  const jsonOptions: string = JSON.stringify(optionsDict);
+  return `${vPythonCommandToUse} ${globalTestInterfacePath} --clicast=${clicastCommandToUse} --mode=rebuild --path=${filePath} --options=${jsonOptions}`;
 }
